requirements.txt
elevenlabs
pyaudio
openai
dotenv
pillow
langchain-community